MIAM-CORE: don't trust pdu->encoding
Encoding information supplied in MIAM CORE header is not reliable.
Some carriers send binary payloads with encoding set to ISO5 or
text payloads with encoding set to binary. In order not to print
binary garbage nor unnecessary hexdumps for text messages, the
decision is now based solely on the message contents. If it contains
only printable characters, it is printed as text. Otherwise
a hexdump is produced.