Emit empty JSON object when node->data is NULL
Sometimes a protocol node does not contain any data fields. A notable
example is MIAM Single Transfer, which is in fact a MIAM CORE Data
PDU without any additional header. Another example is a data type
for which JSON formatting is not yet implemented. We still want to
emit a JSON object with the tag name in such cases, even if the
object will show up as empty.