la_asn1_output(): make ASN.1 dumps optional
In case there is no type-specific ASN.1 formatter defined,
la_asn1_output() produces a raw ASN.1 structure dump for this type.
This is only valid when producing text output and should not
be done for JSON. A new bool argument 'dump_unknown_types' turns
ASN.1 dumps on or off, depending on the context.