acars: parse and store sublabel and message function identifier
This applies to messages with label H1. Sublabel and MFI are
now stripped from message text, which is required for proper
multiblock message reassembly. Note: for service related uplink
messages this is done only if the message text does not
contain SMT header (ARINC 618 4.3). Support for messages with
header might be implemented at a later date, however it is not
critical for reassembly process.