la_reasm_fragment_add(): declare timeouts on duplicate fragments only
This improves handling of ACARS uplink multiblocks with lost fragments.

2019-11-18 22:52:46 u + .SP-LMA T !
2019-11-18 22:52:59 u + .SP-LMA T !
2019-11-18 22:53:05 u + .SP-LMA T !
2019-11-18 22:53:16 u + .SP-LMA T !
2019-11-18 22:53:20 u + .SP-LMA T !
2019-11-18 22:53:21 u + .SP-LMA T !
2019-11-18 22:53:29 u + .SP-LMA T !
2019-11-18 22:53:30 u + .SP-LMA T !
2019-11-18 22:53:39 u + .SP-LMA T !
2019-11-18 22:53:56 u + .SP-LMA T !
2019-11-18 22:54:04 u + .SP-LMA T !
2019-11-18 22:54:14 u + .SP-LMA T !
2019-11-18 22:56:32 u + .SP-LMA V !
2019-11-18 22:56:35 u + .SP-LMA W !
2019-11-18 22:56:37 u + .SP-LMA A !
2019-11-18 22:56:39 u   .SP-LMA B !

Blocks: T and U form message A, block U is lost. Blocks V, W, A, B form
message B. Uplink timeout (90 seconds) expires when V is received.
rt_entry is then deleted. New rt_entry is created on block W. As a
result, message B is reassembled with block V missing. Not good.

This commit causes the timeout to be announced only when the block
received after the timeout is a duplicate of the previous one. If the
sequence number (block ID) is different, then it might be a next
message. Instead of declare a timeout, we ditch the old rt_entry and
create a new one with block V being the first one. Second message is
then reassembled correctly.