Add D-ATIS parsing
D-ATIS is one of the most commonly used message format on the ACARS network, and it is also highly standarized. There are two versions of D-ATIS: the original and obsolete version 1, and the current version 2. This patch supports version 2 only. D-ATIS version 2 uses label A9 / B9 with IMI TI2, as specified in ARINC 623. Related symbols are added.

Signed-off-by: Yuuta Liang <yuuta@yuuta.moe>